Add top-level `Mapping` and `Rules` landing pages
Remove the top-level menu entries `How to Map`, `Mapping Rules`, `Guidelines` and `Auto-Font`.

Add new menu entries `Mapping` and `Rules` as landing pages for the various mapping- and rules-related pages.

The mapping tutorial, rules, guidelines and auto-font tool are now linked on the `Mapping` page. The link to the old mapping tutorial on the forum is replaced with a link to the wiki page, which is more up-to-date and also links back to the forum post.

The masterserver and mapping rules are now linked on the `Rules` page. This is preparation for also adding and linking the Ingame and Community rules.